feature/osk-573-Vyhrady k formularovemu systemu po prvej implementacii (#266)
* Implemented fixes, improvements, and enhancements

* new build for OSK

* Build 2026.0-jakarta-OSK-SNAPSHOT.037 14.07.2026 10:36:49

* fix tests and problems that tests showed

* code review

* tests enhance

* Build 2026.0-jakarta-OSK-SNAPSHOT.038 15.07.2026 8:14:56

* fix quill redundant paragraf

* Allow empty option for DataTableColumnType.OPTIONS

* Add autocomplete type field and support

* Aloow trim value inputed from user

* code review, tests, docs

* Build 2026.0-jakarta-OSK-SNAPSHOT.039 10.08.2026 16:49:32

* new logic, tests and doc edit

* warning in docs

* Merge branch 'main' into feature/osk573-Vyhrady-k-formularovemu-systemu-po-prvej-implementacii

* [osk573] fix: unique ID for fields with same name, form prefix in DOM ids

* [osk573] fix: remove : in email render, fix: compare value with comma

* [osk573] fix: duplicate form in transaction with correct ID mapping to conditions

* [osk573] fix: recaptcha multi form

* [osk573] fix: reCaptchaV3 multi form

* [osk573] fix: invisible captcha, scrollIntoView, better error messages to user

* [osk573] fix: email rendering update codes

* [osk573] feat: check user premission to form before duplicate

* [osk573] fix: migrate to logn for stepId

* [osk573] fix: clean stat when duplicating form

* [osk573] fix: captcha - allow multuple instances, do not store value in session

* [osk573] fix: captcha - allow multuple instances, do not store value in session - v3

* [osk573] fix: captcha - allow multuple instances, do not store value in session - v3

* [osk573] chore: code review, docs

---------

+++ src/main/java/sk/iway/iwcm/components/multistep_form/rest/FormItemsConditionsRestController.java
@@ -53,7 +53,7 @@ public class FormItemsConditionsRestController extends DatatableRestControllerV2
         DatatablePageImpl<FormItemsConditionEntity> page = new DatatablePageImpl<>( super.getAllItemsIncludeSpecSearch(new FormItemsConditionEntity(), pageable) );
 
         String formName = Tools.getStringValue(getRequest().getParameter("formName"), null);
-        int stepId = Tools.getIntValue(getRequest().getParameter("stepId"), -1);
+        long stepId = Tools.getLongValue(getRequest().getParameter("stepId"), -1L);
 
         page.addOptions("itemFormId", multistepFormsService.getAvailableConditionFields(formName, stepId, getProp()), "label", "value", true);
         page.addOptions("operator", OperatorType.getLabelValues(getProp()), "label", "value", true);
